Complete 4.15-4.23

+++ controllers/blogs.js
@@ -1,8 +1,11 @@
+const jwt = require("jsonwebtoken");
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const middleware = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
@@ -15,29 +18,59 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 });
 
-blogsRouter.post("/", async (request, response) => {
-  if (!request.body.title || !request.body.url) {
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
+  const body = request.body;
+  if (body) {
+    console.log(request.token);
+  }
+  if (!body.title || !body.url) {
     return response.status(400).send({ error: "title or url missing" });
   }
-  const body = request.body;
-
+  const user = request.user;
+  console.log(user);
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user.id,
   });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog.id);
+  await user.save();
+
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
-});
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response) => {
+    const blog = await Blog.findById(request.params.id);
 
-blogsRouter.put("/:id", async (request, response) => {
+    if (!blog) {
+      return response
+        .status(404)
+        .json({ error: "Blog not found or already deleted" });
+    } else {
+      console.log("blog found");
+      console.log(request.token);
+    }
+    const user = request.user;
+
+    if (blog.user.toString() !== user._id.toString()) {
+      return response
+        .status(401)
+        .json({ error: "Unauthorized: You are not the creator of this blog" });
+    }
+
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  }
+);
+
+blogsRouter.put("/:id", middleware.userExtractor, async (request, response) => {
   const body = request.body;
 
   const blog = {
